+++ config/estimator_specific/excavator_graph_params.yaml
@@ -1,8 +1,6 @@
 # Sensor Params
 sensor_params:
-  lioOdometryRate: 5 # Output rate of compSLAM
   gnssRate: 20 # Rate of the GNSS
-  useGnssYaw: false  # If true, the yaw is computed from both GNSS positions and added as a factor
 
 # Noise Parameters
 noise_params:
