--- launch/graph_no_urdf.launch
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?> 
-<launch>
-    <!-- Parameters -->
-    <param name="use_sim_time"            value="true"/>
-
-    <!-- Input Topics -->
-    <arg name="imu_topic_name"            default="/microstrain_imu_node/microstrain_imu" />
-    <arg name="lidar_odometry_topic_name" default="/compslam_lio/odom_aft_mapped_to_init_CORRECTED" />
-    <arg name="gnss_left_topic_name"      default="/m545_gnss_left" />
-    <arg name="gnss_right_topic_name"     default="/m545_gnss_right" />
-
-    <!-- Config -->
-    <arg name="extrinsics_from_urdf"      default="false" />
-    <arg name="useLeftGnss"               default="true" />
-    <arg name="useRightGnss"               default="true" />
-    <arg name="useLioOdometry"                 default="true" />
-
-    <!-- Description -->
-    <arg name="description_name"          default="/heap_description" />
-
-    <!-- Config Files -->
-    <arg name="core_graph_config_file"         default="$(find leica_position_graph)/config/core/core_graph_config.yaml"/>
-    <arg name="core_graph_parameter_file"      default="$(find leica_position_graph)/config/core/core_graph_params.yaml"/>
-    <arg name="core_extrinsic_file"            default="$(find leica_position_graph)/config/core/core_extrinsic_params.yaml"/>
-    <arg name="excavator_graph_parameter_file" default="$(find leica_position_graph)/config/excavator_specific/excavator_graph_params.yaml"/>
-    <arg name="excavator_extrinsic_file"       default="$(find leica_position_graph)/config/excavator_specific/excavator_extrinsic_params.yaml"/>
-    <arg name="excavator_gnss_parameter_file"  default="$(find leica_position_graph)/config/excavator_specific/excavator_gnss_params.yaml"/>
-
-    <!-- Node -->
-    <node pkg="leica_position_graph" type="leica_position_graph_node" name="holistic_graph_node" output="screen" > <!-- launch-prefix="gdb -args"-->
-        <!-- Launch Parameters -->
-        <param name="launch/descriptionName" type="string" value="$(arg description_name)" />
-        <param name="launch/useLeftGnss" type="bool" value="$(arg useLeftGnss)" />
-        <param name="launch/useRightGnss" type="bool" value="$(arg useRightGnss)" />
-        <param name="launch/useLioOdometry" type="bool" value="$(arg useLioOdometry)" />
-        <!-- Parameter files -->
-        <rosparam command="load" file="$(arg core_graph_config_file)" />
-        <rosparam command="load" file="$(arg core_graph_parameter_file)" />
-        <rosparam command="load" file="$(arg core_extrinsic_file)" />
-        <rosparam command="load" file="$(arg excavator_graph_parameter_file)" />
-        <rosparam command="load" file="$(arg excavator_extrinsic_file)" />
-        <rosparam command="load" file="$(arg excavator_gnss_parameter_file)" />
-        <!-- Remapping of topics -->
-        <remap from="/imu_topic"            to="$(arg imu_topic_name)" />
-        <remap from="/lidar_odometry_topic" to="$(arg lidar_odometry_topic_name)" />
-        <remap from="/gnss_topic_1"         to="$(arg gnss_left_topic_name)" />
-        <remap from="/gnss_topic_2"         to="$(arg gnss_right_topic_name)" />
-    </node>
-
-    <!-- RVIZ for Visualization -->
-    <node type="rviz" name="rviz" pkg="rviz" args="-d $(find leica_position_graph)/rviz/rviz_imuOdom.rviz" />
-
-</launch>
