+++ src/graph_node.cpp
@@ -8,12 +8,12 @@ Please see the LICENSE file that has been included as part of this package.
 // ROS
 #include <ros/ros.h>
 // Local packages
-#include "excavator_holistic_graph/ExcavatorEstimator.h"
+#include "leica_position_graph/PositionEstimator.h"
 
 // Main node entry point
 int main(int argc, char** argv) {
   // ROS related
-  ros::init(argc, argv, "laserOdometry");
+  ros::init(argc, argv, "leica_position_graph_node");
   std::shared_ptr<ros::NodeHandle> privateNodePtr = std::make_shared<ros::NodeHandle>("~");
   /// Do multi-threaded spinner
   ros::MultiThreadedSpinner spinner(4);
